# docs/tee-api-openapi.yaml
openapi: 3.1.0
...
info:
  title: Nyx TEE API
  version: 2.0.0-draft
  description: |
    Public + authenticated REST and WebSocket API for the Nyx dark pool
    TEE matching layer. Pairs with the v2 on-chain custody program
    (vault, program ID `C63vKvysCzX55PKraas4Wc22ijqjGJQdPC1mrzCFVWZx`).
  contact:
    name: Nyx engineering
    url: https://github.com/Nyx-Privacy/nyx
  license:
    name: Apache-2.0
servers:
  - url: https://api.nyx.example.com/api/v1
    description: Mainnet (TBD — Phala dstack)
  - url: https://api.devnet.nyx.example.com/api/v1
    description: Devnet (TBD — Phala dstack devnet)

# ─────────────────────────────────────────────────────────────────────────────
# Tags — grouped for docs rendering
# ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived (≤ 1h) bearer token from POST /auth/token.

  # ───── Shared schemas ──────────────────────────────────────────────────────
  schemas:
...
    AttestationQuote:
      type: object
      description: |
        Intel TDX attestation quote + event log + VM config.
        Shape matches what the dstack-sdk `getQuote()` returns plus
        a few convenience fields the SDK's verifyTeeAttestation()
        helper needs. Clients run the verification chain documented
        in docs/tee-attestation-flow.md §4.
      required: [quote, event_log, report_data, tee_pubkey]
      properties:
        quote:
          type: string
          description: Hex-encoded TDX quote (DCAP format).
        event_log:
          type: string
          description: |
            The dstack event log as a JSON STRING (an array of measured
            events — NOT hex-encoded). A client replays it against the
            DCAP-verified quote's RTMR3 to bind the recorded compose-hash
            + instance-id + key-provider events
            (packages/sdk/src/tee/verify-core.ts).
        report_data:
          type: string
          description: |
            64-byte report_data field embedded in the quote, hex.
            Layout when callers pass `?reportData=<nonce>`:
              report_data[0..32] = caller-supplied nonce
              report_data[32..64] = SHA-256(tee_pubkey)  (binding)
        vm_config:
          type: object
          description: |
            OPTIONAL. VM hardware configuration the quote attests to
            (dstack `getQuote` may include it). The nyx-tee handler does
            not currently return this field — do not rely on it.
          properties:
            cpu_count: { type: integer }
            memory_mb: { type: integer }
            os_image_hash: { type: string }
        tee_pubkey:
          type: string
          description: |
            Base58 Ed25519 pubkey the TEE signs MatchResultPayload
            canonical hashes with. Equal to on-chain
            `vault_config.tee_pubkey` after the most recent
            set_tee_pubkey multisig rotation.
        issued_at:
          type: string
          format: date-time
        # The MRTD / RTMR0-3 fields are present inside the binary
        # quote and are exposed by dstack-verifier when verification
        # runs. They're not duplicated here as separate response
        # fields — clients parse them out of `quote` via dcap-qvl
        # (or by hitting Phala's /api/v1/attestations/verify endpoint).
...
    TransparencySnapshot:
      type: object
      required: [reserves, tee, stats]
      description: |
        Public, unauthenticated. Anyone can verify
        (a) vault token balances cover liabilities + (b) the TEE
        attestation is current + (c) aggregate operations are
        healthy.
      properties:
        reserves:
          type: object
          required: [merkle_root, leaf_count, per_mint]
          properties:
            merkle_root:
              type: string
              description: |
                Shard-0 root for back-compat. Under tree-sharding each
                shard has its own root; read per-shard via
                `/tree/root?tree_id=`.
            leaf_count:
              type: integer
              description: Total leaves SUMMED across all `num_trees` shards.
            per_mint:
              type: array
              items:
                type: object
                required: [mint, outstanding, vault_balance, stale]
                properties:
                  mint: { type: string }
                  outstanding:
                    type: string
                    description: |
                      Sum of un-spent note amounts for this mint
                      (== `OutstandingMint.outstanding`).
                  vault_balance:
                    type: string
                    description: |
                      Actual SPL balance in the vault's PDA for
                      this mint. MUST be >= outstanding (the v2
                      solvency invariant). If not, something is
                      wrong on-chain.
                  stale:
                    type: boolean
                    description: |
                      true if either on-chain read was DEGRADED (RPC
                      error or malformed account); outstanding /
                      vault_balance are then "unknown", not a real 0 —
                      ignore the numbers for a solvency check.
        tee:
          type: object
          description: |
            Engine attestation IDENTITY — enough to tie this response to
            a measured image. The full challenge-response quote is at
            GET /attestation.
          required: [app_id, compose_hash, mrtd, signer_pubkey]
          properties:
            app_id: { type: string }
            compose_hash: { type: string }
            mrtd: { type: string }
            signer_pubkey: { type: string, description: TEE Ed25519 signer (base58). }
        stats:
          type: object
          description: |
            Settle-scheduler counters tracked today. The 24h windows +
            finality timing (avg_settlement_finality_ms, settles_24h, …)
            are a later addition once the scheduler records timestamps.
          required: [batches, jobs]
          properties:
            batches: { type: integer, description: Settle batches tracked. }
            jobs: { type: integer, description: Per-match settle jobs tracked. }
...
paths:
...
  /attestation:
    get:
      tags: [attestation]
      summary: Fresh TEE attestation quote (challenge-response).
      description: |
        Unauthenticated. Clients MUST verify this quote before
        sending order intent. Full verification chain documented
        in docs/tee-attestation-flow.md §4.

        The caller supplies a fresh nonce via reportData; the TEE
        generates a quote with reportData populated as:
            report_data[0..32]  = caller's nonce
            report_data[32..64] = SHA-256(tee_pubkey)

        Verification steps (handled by the SDK's
        verifyTeeAttestation() helper):
          1. Verify Intel TCB cert chain (locally via dcap-qvl or
             via Phala's POST /api/v1/attestations/verify).
          2. Replay RTMR3 event log → confirm rt_mr3 in quote.
          3. Confirm compose-hash event in RTMR3 == the SDK's
             baked-in EXPECTED_COMPOSE_HASH constant.
          4. Confirm key-provider event == known-trusted KMS ID.
          5. Confirm reportData starts with the supplied nonce
             (replay protection) AND ends with SHA-256(tee_pubkey).
          6. Optionally: check CAA records + CT logs (paranoid).
      parameters:
        - name: reportData
          in: query
          required: false
          schema: { type: string }
          description: |
            32-byte nonce in hex. If omitted, the TEE generates a
            quote with all-zero reportData[0..32]. Always supply a
            fresh nonce for live verification (replay protection).
      responses:
        '200':
          description: Current quote.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttestationQuote' }
...
  /transparency:
    get:
      tags: [transparency]
      summary: |
        Public proof-of-reserves + engine attestation + aggregate
        stats. Unauthenticated. Anyone can verify.
      parameters:
        - name: epoch
          in: query
          schema: { type: integer }
          description: |
            Optional. Returns the snapshot for a specific TEE
            epoch (rotation generation). Default is current.
      responses:
        '200':
          description: Transparency snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransparencySnapshot' }

  # ───── settlement ──────────────────────────────────────────────────────────
...